+++ iphash/iphashtest4.c
@@ -108,7 +108,7 @@ static void batch_timer_fn(
       {
         tokens = initial_tokens;
       }
-      e->tokens = tokens;
+      e->tokens = (uint16_t)tokens;
     }
   }
   else
@@ -199,7 +199,7 @@ int main(int argc, char **argv)
       {
         tokens = initial_tokens;
       }
-      e->tokens = tokens;
+      e->tokens = (uint16_t)tokens;
     }
   }
   else
@@ -230,7 +230,7 @@ int main(int argc, char **argv)
       {
         tokens = initial_tokens;
       }
-      e->tokens = tokens;
+      e->tokens = (uint16_t)tokens;
     }
   }
   else
