+++ ldp/ldpreplay.c
@@ -88,7 +88,7 @@ int main(int argc, char **argv)
       exit(1);
     }
   }
-  sz = DYNARR_SIZE(&ar);
+  sz = (int)DYNARR_SIZE(&ar);
   last_time64 = gettime64();
   for (;;)
   {
@@ -97,7 +97,7 @@ int main(int argc, char **argv)
       pkts++;
       bytes += DYNARR_GET(&ar, i)->len;
       pkt_tbl[cnt].data = DYNARR_GET(&ar, i)->data;
-      pkt_tbl[cnt].sz = DYNARR_GET(&ar, i)->len;
+      pkt_tbl[cnt].sz = (uint32_t)DYNARR_GET(&ar, i)->len;
       cnt++;
       if (cnt == sizeof(pkt_tbl)/sizeof(*pkt_tbl))
       {
@@ -110,8 +110,8 @@ int main(int argc, char **argv)
         uint64_t time64 = gettime64();
         uint64_t pdiff = pkts - last_pkts;
         uint64_t bdiff = bytes - last_bytes;
-        double tdiff = time64 - last_time64;
-        printf("%g MPPS %g Gbps\n", pdiff/tdiff, bdiff*8/1000/tdiff);
+        double tdiff = (double)(time64 - last_time64);
+        printf("%g MPPS %g Gbps\n", (double)pdiff/tdiff, (double)bdiff*8/1000/tdiff);
         last_pkts = pkts;
         last_bytes = bytes;
         last_time64 = time64;
