+++ ldp/ldptunnel.c
@@ -99,10 +99,10 @@ int main(int argc, char **argv)
       uint64_t dlpdiff = dlpkts - last_dlpkts;
       uint64_t ulbdiff = ulbytes - last_ulbytes;
       uint64_t dlbdiff = dlbytes - last_dlbytes;
-      double tdiff = time64 - last_time64;
+      double tdiff = (double)(time64 - last_time64);
       printf("%g MPPS %g Gbps ul, %g MPPS %g Gbps dl\n",
-             ulpdiff/tdiff, ulbdiff*8/1000/tdiff,
-             dlpdiff/tdiff, dlbdiff*8/1000/tdiff);
+             (double)ulpdiff/tdiff, (double)ulbdiff*8/1000/tdiff,
+             (double)dlpdiff/tdiff, (double)dlbdiff*8/1000/tdiff);
       last_time64 = time64;
       last_ulpkts = ulpkts;
       last_dlpkts = dlpkts;
@@ -124,7 +124,7 @@ int main(int argc, char **argv)
         ulbytes += pkts[i].sz;
 
         ip = ether_payload((void*)tunnelhdrs[i]);
-        ip_set_total_len(ip, pkts[i].sz+4+20);
+        ip_set_total_len(ip, (uint16_t)(pkts[i].sz+4+20)); // XXX int overflow
         ip_set_hdr_cksum_calc(ip, 20);
         iovs[i][0].iov_base = tunnelhdrs[i];
         iovs[i][0].iov_len = 14+20+4;
